requirements.txt
langchain-core==0.2.43
langchain-openai==0.1.23
pydantic==2.9.2
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
plotly==5.24.1
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.2
